requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
streamlit==1.28.1
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
